pyproject.toml
[tool.poetry]
name = "grpc-sample"
version = "0.1.0"
description = ""
authors = ["kishimoto <drehbleistift@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
grpcio-tools = "^1.38.1"
grpcio-reflection = "^1.38.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
